TypeScript convert ResourceTable. Removed footer.
